refactor: move app_exception_handler to example directory
