+++ example/main.dart
@@ -1,5 +1,5 @@
 import 'package:vania/vania.dart';
-import 'package:vania/src/exception/app_exception_handler.dart';
+import 'app_exception_handler.dart';
 
 void main() async {
   Application().initialize(
